eurorate-predictor/requirements.txt
zenml[server]
pandas 
xgboost
google-cloud-bigquery
pyarrow
gradio
db-dtypes